# swagger.yaml
openapi: 3.0.0
info:
  title: Animal Farm API
  description: |
    A RESTful API for managing farm animals with OOP design patterns.
    
    ## Features
    - Create and manage birds and dogs
    - Set and perform duties and actions
    - Manage animal states (hunger, sleepiness)
    - Runtime action updates
    - Health monitoring and metrics
    
    ## Design Patterns
    - Template Method Pattern
    - Strategy Pattern
    - Factory Pattern
    - Repository Pattern
    - State Pattern
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: ISC

servers:
  - url: http://localhost:3000
    description: Development server

tags:
  - name: Health
    description: Health check and metrics endpoints
  - name: Animals
    description: Animal management operations
  - name: Actions
    description: Animal action management
  - name: States
    description: Animal state management

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns the health status of the API
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  uptime:
                    type: string
                    example: 3600s
                  timestamp:
                    type: string
                    format: date-time
                  animalsCount:
                    type: integer
                    example: 5

  /metrics:
    get:
      tags:
        - Health
      summary: Get API metrics
      description: Returns business metrics about the farm
      responses:
        '200':
          description: Metrics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalAnimals:
                    type: integer
                  birds:
                    type: integer
                  dogs:
                    type: integer
                  hungryAnimals:
                    type: integer
                  sleepyAnimals:
                    type: integer
                  animalsWithDuties:
                    type: integer
                  animalsWithActions:
                    type: integer

  /animals:
    get:
      tags:
        - Animals
      summary: List all animals
      description: Returns a list of all animals on the farm
      responses:
        '200':
          description: List of animals
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  animals:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnimalStatus'

  /animals/bird:
    post:
      tags:
        - Animals
      summary: Create a new bird
      description: Creates a new bird with specified abilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Rooster
                canFly:
                  type: boolean
                  default: false
                  example: false
                canCrow:
                  type: boolean
                  default: false
                  example: true
      responses:
        '201':
          description: Bird created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  animal:
                    $ref: '#/components/schemas/AnimalStatus'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /animals/dog:
    post:
      tags:
        - Animals
      summary: Create a new dog
      description: Creates a new dog with specified abilities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Rex
                canBark:
                  type: boolean
                  default: false
                  example: true
                canChase:
                  type: boolean
                  default: false
                  example: true
      responses:
        '201':
          description: Dog created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  animal:
                    $ref: '#/components/schemas/AnimalStatus'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /animals/{name}:
    get:
      tags:
        - Animals
      summary: Get animal details
      description: Returns detailed information about a specific animal
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          example: Rex
      responses:
        '200':
          description: Animal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimalStatus'
        '404':
          description: Animal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Animals
      summary: Delete an animal
      description: Removes an animal from the farm
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Animal deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Animal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /animals/{name}/duty:
    put:
      tags:
        - Actions
      summary: Set animal duty
      description: Assigns a duty to an animal (runtime update)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - duty
              properties:
                duty:
                  type: string
                  example: guard the farm
      responses:
        '200':
          description: Duty set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    $ref: '#/components/schemas/AnimalStatus'
        '404':
          description: Animal not found

  /animals/{name}/perform-duty:
    post:
      tags:
        - Actions
      summary: Perform duty
      description: Makes the animal perform its assigned duty
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Duty performed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    $ref: '#/components/schemas/AnimalStatus'
        '400':
          description: Cannot perform duty (no duty assigned, hungry, or sleepy)

  /animals/{name}/action:
    put:
      tags:
        - Actions
      summary: Set animal action
      description: Assigns an action to an animal (runtime update)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: [fly, crow, bark, chase]
                  example: bark
      responses:
        '200':
          description: Action set successfully
        '400':
          description: Invalid action for this animal type

  /animals/{name}/perform-action:
    post:
      tags:
        - Actions
      summary: Perform action
      description: Makes the animal perform its assigned action
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Action performed successfully
        '400':
          description: Cannot perform action

  /animals/{name}/hungry:
    put:
      tags:
        - States
      summary: Set hungry state
      description: Updates the animal's hunger state
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hungry
              properties:
                hungry:
                  type: boolean
                  example: true
      responses:
        '200':
          description: State updated successfully

  /animals/{name}/sleepy:
    put:
      tags:
        - States
      summary: Set sleepy state
      description: Updates the animal's sleepiness state
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sleepy
              properties:
                sleepy:
                  type: boolean
                  example: true
      responses:
        '200':
          description: State updated successfully

  /animals/{name}/eat:
    post:
      tags:
        - States
      summary: Make animal eat
      description: Feeds the animal (only works if hungry)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Animal ate successfully

  /animals/{name}/sleep:
    post:
      tags:
        - States
      summary: Make animal sleep
      description: Lets the animal sleep (only works if sleepy)
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Animal slept successfully

components:
  schemas:
    AnimalStatus:
      type: object
      properties:
        name:
          type: string
          example: Rex
        type:
          type: string
          enum: [Bird, Dog]
          example: Dog
        legs:
          type: integer
          example: 4
        isHungry:
          type: boolean
          example: false
        isSleepy:
          type: boolean
          example: false
        currentDuty:
          type: string
          nullable: true
          example: guard the farm
        currentAction:
          type: string
          nullable: true
          example: bark
        availableActions:
          type: array
          items:
            type: string
          example: [bark, chase]
    
    Error:
      type: object
      properties:
        error:
          type: string
          example: Animal not found